==> 1_DPP4-External-Data-Collection & Feature Extraction.ipynb <==
ROOT_DIR ='/data/user/home/mhossai5/DPP-New2025'
!git clone https://github.com/brendaferrari/AutoPaDELPy.git -l {ROOT_DIR}/utils

import os
import sys
sys.path.insert(0,os.path.join(
    ROOT_DIR,'utils'
))
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

# Data Collection

df_external_set = pd.read_excel(os.path.join(
    ROOT_DIR,'data/bioactivities_Tue_May_06_2025.xlsx'
))
df_external_set.head()

df_external_set.info()

df_external_set= df_external_set[(
    df_external_set['Target Pref Name'].isin(['DIPEPTIDYL PEPTIDASE IV']) 
    & 
    df_external_set['Gene Names'].isin(['DPP4'])
    &
    df_external_set['End Point Standard Type'].isin(['IC50'])
    &
    df_external_set['End Point Standard Units'].isin(['NM'])
)]
df_external_set = df_external_set.drop_duplicates(['Standard inchi key']).reset_index(drop = True)

df_external_set.shape

# ref: https://drugtargetcommons.fimm.fi/bioactivities?id=DTCT0024079&category=Target&name=DPP4
import pubchempy as pcp
import pandas as pd
from tqdm.auto import tqdm

# Function using pubchempy
def inchikey_to_smiles_pubchempy(inchikey):
    compounds = pcp.get_compounds(inchikey, 'inchikey')
    if compounds and compounds[0].canonical_smiles:
        return compounds[0].canonical_smiles
    return None

# Add SMILES column
df_external_set['SMILES'] = None
for idx,row in tqdm(df_external_set.iterrows(),total = len(df_external_set)):
    inchikey = row['Standard inchi key']
    if pd.isna(df_external_set.at[idx,'SMILES']):
        try:
            df_external_set.at[idx,'SMILES'] = inchikey_to_smiles_pubchempy(inchikey)
        except:
            df_external_set.at[idx,'SMILES'] = None

df_external_set = df_external_set[['DTC Activity ID','SMILES','End Point Standard Value','End Point Standard Type','End Point Standard Units']]
df_external_set.rename(columns = {
    'DTC Activity ID':'id',
    'SMILES':'smiles',
    'End Point Standard Value':'standard_value',
    'End Point Standard Type':'standard_type',
    'End Point Standard Units':'standard_units'
},inplace = True)
df_external_set['target_organism'] = 'Homo sapiens'
df_external_set['target_pref_name'] = 'Dipeptidyl peptidase IV'
df_external_set = df_external_set[['id','smiles','standard_value','standard_type','standard_units','target_organism','target_pref_name']]
df_external_set.dropna(inplace = True)
df_external_set

df_external_set.to_csv(os.path.join(ROOT_DIR,'data/External_DTC_DPP4-05-05-2025_Processed.csv'),index = False)


# Feature Extraction

import pandas as pd
import numpy as np
from tqdm.auto import tqdm
tqdm.pandas()
df_external_set = pd.read_csv(os.path.join(ROOT_DIR,'data/External_DTC_DPP4-05-05-2025_Processed.csv'))
df_train = pd.read_parquet(os.path.join(ROOT_DIR,'data/dpp4-26-03-25-feat.parquet'))

# ## IC50 Conversion
def calculate_pIC50(ic50_nM):
    ic50_nM = pd.to_numeric(ic50_nM, errors='coerce')  # Convert to numeric, handle errors
    ic50_M = ic50_nM * 1e-9  # Convert to molar concentration
    return -np.log10(ic50_M)  # Compute pIC50
def classify_pic50(value):
    if value > 7.5:
        return 'active'
    elif 6 < value <= 7.5:
        return 'grey'
    elif value <= 6:
        return 'inactive'
# Save the fitted LabelEncoder to a file
import pickle

with open(os.path.join(ROOT_DIR,'data/label_encoder.pkl'), 'rb') as file:
    le = pickle.load(file)
    
with open(os.path.join(ROOT_DIR,'data/ltn_std_scaler.pkl'), 'rb') as file:
    scaler = pickle.load(file)
    
df_external_set['pIC50'] = df_external_set['standard_value'].apply(calculate_pIC50)
# Apply the classification to the DataFrame
df_external_set['label'] = df_external_set['pIC50'].apply(classify_pic50)
df_external_set = df_external_set[~df_external_set['label'].isin(['grey'])]
df_external_set['target'] = le.transform(df_external_set['label'])

df_external_set = df_external_set[~df_external_set['smiles'].progress_apply(lambda x: x in df_train['smiles'].tolist())].reset_index(drop = True)
df_external_set

from functions.fingerprint_functions import FingerprintFunctions

feat_path = os.path.join(ROOT_DIR,'data/External_DTC_DPP4-CDKextended.csv')
if not os.path.exists(feat_path):
    os.chdir(os.path.join(ROOT_DIR,'utils'))
    smi_path =  'dataset.smi'
    df_external_set[['smiles','id']].to_csv(
       smi_path, sep='\t', index=False, header=False,
    )
    FingerprintFunctions().do_CDKextended(smi_path)
    os.rename(
        os.path.join(ROOT_DIR,'utils/CDKextended.csv'),
        feat_path,
    )
    os.remove(smi_path)
    os.chdir(os.path.join(ROOT_DIR,'notebook'))

df_external_set['CDKextended'] = pd.read_csv(feat_path).drop('Name',axis = 1).values.tolist()

import rdkit
from rdkit import Chem
from rdkit.Chem import AllChem
from tqdm.auto import tqdm
tqdm.pandas()
import numpy as np
from rdkit import rdBase
rdBase.DisableLog('rdApp.warning')

def smiles_to_fp(smiles, radius=2, num_bits=1024):
    mol = Chem.MolFromSmiles(smiles)
    if mol is None:
        return list(np.zeros(num_bits))
    return list(AllChem.GetMorganFingerprintAsBitVect(mol, radius, num_bits))

df_external_set['ECFP'] = df_external_set['smiles'].progress_apply(lambda x: smiles_to_fp(x))
df_external_set['ECFP_2048'] = df_external_set['smiles'].progress_apply(lambda x: smiles_to_fp(x,num_bits = 2048))
df_external_set['ECFP_512'] = df_external_set['smiles'].progress_apply(lambda x: smiles_to_fp(x,num_bits = 512))

df_external_set.to_parquet(os.path.join(ROOT_DIR,'data/dpp4-ext-26-03-25-feat.parquet'),index = False)
