==> 2_DPP4-Features-Exatraction.ipynb <==
ROOT_DIR ='/data/user/home/mhossai5/DPP-New2025'
!git clone https://github.com/brendaferrari/AutoPaDELPy.git -l {ROOT_DIR}/utils

import os
import sys
sys.path.insert(0,os.path.join(
    ROOT_DIR,'utils'
))
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

df_final = pd.read_csv(os.path.join(
    ROOT_DIR,'data/dpp4-26-03-25.csv'
))
df_final.head()

# PaDEL Descriptors

from functions.fingerprint_functions import FingerprintFunctions

if not os.path.exists(os.path.join(ROOT_DIR,'data/CDKextended-dpp4.csv')):
    os.chdir(os.path.join(ROOT_DIR,'utils'))
    smi_path =  'dataset.smi'
    df_final[['smiles','id']].to_csv(
       smi_path, sep='\t', index=False, header=False,
    )
    FingerprintFunctions().do_CDKextended(smi_path)
    os.rename(
        os.path.join(ROOT_DIR,'utils/CDKextended.csv'),
        os.path.join(ROOT_DIR,'data/CDKextended-dpp4.csv'),
    )
    os.remove(smi_path)
    os.chdir(os.path.join(ROOT_DIR,'notebook'))

df_final['CDKextended'] = pd.read_csv( os.path.join(ROOT_DIR,'data/CDKextended-dpp4.csv')).drop('Name',axis = 1).values.tolist()

# Morgan 

import rdkit
from rdkit import Chem
from rdkit.Chem import AllChem
from tqdm.auto import tqdm
tqdm.pandas()
import numpy as np
from rdkit import rdBase
rdBase.DisableLog('rdApp.warning')

def smiles_to_fp(smiles, radius=2, num_bits=1024):
    mol = Chem.MolFromSmiles(smiles)
    if mol is None:
        return list(np.zeros(num_bits))
    return list(AllChem.GetMorganFingerprintAsBitVect(mol, radius, num_bits))

df_final['ECFP'] = df_final['smiles'].progress_apply(lambda x: smiles_to_fp(x))
df_final['ECFP_2048'] = df_final['smiles'].progress_apply(lambda x: smiles_to_fp(x,num_bits = 2048))
df_final['ECFP_512'] = df_final['smiles'].progress_apply(lambda x: smiles_to_fp(x,num_bits = 512))

df_final.head()

import torch
import numpy as np
import pandas as pd
from tqdm import tqdm
from transformers import AutoModel, AutoTokenizer

# Load Llama model and tokenizer
model_name = "meta-llama/Llama-3.2-1B"
device = "cuda" if torch.cuda.is_available() else "cpu"

tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModel.from_pretrained(model_name).to(device)
tokenizer.pad_token = tokenizer.eos_token
# Function to extract embeddings
def smiles_to_embedding(smiles):
    """Convert SMILES string to Llama embeddings."""
    inputs = tokenizer(smiles, padding=True, truncation=True, return_tensors="pt").to(device)
    
    with torch.no_grad():
        outputs = model(**inputs)
        embedding = outputs.last_hidden_state.mean(dim=1).squeeze().cpu().numpy().squeeze()  # Mean pooling

    return embedding

df_final['Llama3-2_Embedding'] = df_final['smiles'].progress_apply(lambda x: smiles_to_embedding(x))

# Print sample embeddings
df_final.head()

# Load ChemBERT2 model and tokenizer
model_name = "seyonec/ChemBERTa_zinc250k_v2_40k"
device = "cuda" if torch.cuda.is_available() else "cpu"

tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModel.from_pretrained(model_name).to(device)

# Function to extract embeddings
def smiles_to_embedding(smiles):
    """Convert SMILES string to ChemBERT2 embeddings."""
    inputs = tokenizer(smiles, padding=True, truncation=True, return_tensors="pt").to(device)
    
    with torch.no_grad():
        outputs = model(**inputs)
        embedding = outputs.last_hidden_state.mean(dim=1).squeeze().cpu().numpy().squeeze()  # Mean pooling

    return embedding

df_final['ChemBERT2_Embedding'] = df_final['smiles'].progress_apply(lambda x: smiles_to_embedding(x))

# Print sample embeddings
df_final.head()

import pandas as pd
import numpy as np
from rdkit import Chem
from rdkit.Chem import Descriptors, Crippen, Lipinski

 # Function to compute physicochemical properties
def compute_physicochemical_properties(smiles):
    mol = Chem.MolFromSmiles(smiles)
    if mol is None:
        return {
            "MolWt": np.nan, "LogP": np.nan, "TPSA": np.nan,
            "HBD": np.nan, "HBA": np.nan, "RotatableBonds": np.nan
        }
    
    return {
        "MolWt": Descriptors.MolWt(mol),               # Molecular Weight
        "LogP": Crippen.MolLogP(mol),                  # LogP (Hydrophobicity)
        "TPSA": Descriptors.TPSA(mol),                 # Topological Polar Surface Area
        "HBD": Lipinski.NumHDonors(mol),               # Hydrogen Bond Donors
        "HBA": Lipinski.NumHAcceptors(mol),            # Hydrogen Bond Acceptors
        "RotatableBonds": Lipinski.NumRotatableBonds(mol)  # Rotatable Bonds
    }

# Apply the function and expand results into columns
df_final['Physiochemical'] = df_final['smiles'].apply(lambda x: np.nan_to_num(list(compute_physicochemical_properties(x).values())))

df_final.to_parquet(os.path.join(
    ROOT_DIR,'data/dpp4-26-03-25-feat.parquet'
),index = False)

df_final.head()

df_final['ChemBERT2_Embedding'][0].shape
