==> 3_DPP4-LTN-Training-CDKExetended-ECFP-HP.ipynb <==
!wget https://github.com/logictensornetworks/logictensornetworks/raw/master/examples/multiclass_classification/commons.py
!pip install ltn "tensorflow==2.15.*" -q
!pip install tensorflow[and-cuda]==2.15.* --q
!pip install keras_tuner -q

!mkdir -p data
!gdown 1SGiYOyuSiirueZR3F6K0d7aHcPT43PYw -O ./data/dpp4-26-03-25-feat.parquet
!ls -l data/dpp4-26-03-25-feat.parquet
!mkdir -p results
!mkdir -p weights

import os
import ltn
import numpy as np
import pandas as pd
import tensorflow as tf
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split

from sklearn.metrics import (
    accuracy_score,
    f1_score,
    roc_auc_score,
    roc_curve,
    matthews_corrcoef,
    precision_score,
    recall_score,
    confusion_matrix,
    classification_report,
    auc,
    ConfusionMatrixDisplay,
)

import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.auto import tqdm
from dataclasses import dataclass
tqdm.pandas()
import matplotlib.pyplot as plt
import keras_tuner as kt
from tensorflow import keras

tf.config.list_physical_devices('GPU')

ROOT_DIR = '/content'

# Data & Model training arguments
SEED = 42
LR = 0.00001
BATCH_SIZE = 128
FEAT_NAME = [
    'CDKextended',
    'ECFP',
    'ECFP_512',
    'ECFP_2048',
    # 'ChemBERT2_Embedding',
    # 'Llama3-2_Embedding',
    # 'MolWt',
    # 'LogP',
    # 'TPSA',
    # 'HBD',
    # 'HBA',
    # 'RotatableBonds',
]
MODEL_NAME = 'LTN'

TRAIN_HISTORY_PATH = os.path.join(ROOT_DIR, f'results/{"-".join(FEAT_NAME)}-{MODEL_NAME}.csv')
os.makedirs(os.path.dirname(TRAIN_HISTORY_PATH), exist_ok=True)

import pandas as pd
import os

# Define the correct path
file_path = os.path.join(ROOT_DIR, "data/dpp4-26-03-25-feat.parquet")

# Read the Parquet file
df_final = pd.read_parquet(file_path)

df_final.head()

df_final.shape

# Operators and axioms

Not = ltn.Wrapper_Connective(ltn.fuzzy_ops.Not_Std())
And = ltn.Wrapper_Connective(ltn.fuzzy_ops.And_Prod())
Or = ltn.Wrapper_Connective(ltn.fuzzy_ops.Or_ProbSum())
Implies = ltn.Wrapper_Connective(ltn.fuzzy_ops.Implies_Reichenbach())
Forall = ltn.Wrapper_Quantifier(ltn.fuzzy_ops.Aggreg_pMeanError(p=2),semantics="forall")

class MLP(tf.keras.Model):
    """Model that returns logits."""
    def __init__(self, n_classes,units_values= [16,16,8], activation ='elu'):
        super(MLP, self).__init__()
        self.denses = [tf.keras.layers.Dense(uv, activation=activation) for uv in units_values]
        self.dense_class = tf.keras.layers.Dense(n_classes)
        self.dropout = tf.keras.layers.Dropout(0.25)

    def call(self, inputs, training=False):
        x = inputs
        for dense in self.denses:
            x = dense(x)
            x = self.dropout(x, training=training)

        return self.dense_class(x)

def build_model(hp):

    units_value1 = hp.Choice("units_value1",[
        768,384,192,
        ])
    units_value2 = hp.Choice("units_value2",[
        384,192,
        ])
    units_value3 = hp.Choice("units_value3",[
        192,96,
        ])

    activations = hp.Choice("activation", ["relu", 'elu'])
    model = MLP(
        n_classes=2,
        units_values=[units_value1,units_value2,units_value3],
        activation=activations,
    )
    predicate = LTNModel(model)
    predicate.compile(
        optimizer=keras.optimizers.Adam(
            hp.Choice("learning_rate",[1e-5, 1e-4, 1e-3])
        ),
    )
    return predicate

formula_aggregator = ltn.Wrapper_Formula_Aggregator(ltn.fuzzy_ops.Aggreg_pMeanError(p=4))

# Creating Object for Training
class LTNModel(tf.keras.Model):
    def __init__(self,input_model):
        super(LTNModel, self).__init__()

        self.model = input_model # accuracy prediction
        self.p = ltn.Predicate(ltn.utils.LogitsToPredicateModel(self.model,single_label=True)) # satification, main ltn
        # self.p = ltn.Predicate.FromLogits(self.model, activation_function="softmax", with_class_indexing=True)
        self.train_sat = tf.keras.metrics.Mean(name='train_axiom_sat')
        self.test_sat = tf.keras.metrics.Mean(name='test_axiom_sat')
        self.train_accu = tf.keras.metrics.Accuracy(name='train_accuracy')
        self.test_accu = tf.keras.metrics.Accuracy(name='test_accuracy')

    @tf.function
    def axioms(self,features , labels,p,training = False):

        class_A = ltn.Constant(0, trainable=False)
        class_B = ltn.Constant(1, trainable=False)

        x_A = ltn.Variable("x_A",features[labels==0])
        x_B = ltn.Variable("x_B",features[labels==1])

        axioms = [
            Forall(x_A,p([x_A,class_A],training = training)),
            Forall(x_B,p([x_B,class_B],training = training)),
            ]
        for i in range(len(axioms)):
            # find nan float values and replace them with 0
            if tf.reduce_any(tf.math.is_nan(axioms[i].tensor)):
                axioms[i].tensor = tf.constant(0.0)

        sat_level = formula_aggregator(axioms).tensor
        return sat_level


    @tf.function
    def train_step(self,data_train):
        # sat and update
        x,y = data_train
        with tf.GradientTape(persistent = False) as tape:
            self.sat = self.axioms(x, y,self.p,training = True)
            self.loss_value = 1. -self.sat # Learning

        grads = tape.gradient(self.loss_value, self.p.trainable_variables)
        self.optimizer.apply_gradients(zip(grads, self.p.trainable_variables))
        pred = tf.argmax(tf.nn.softmax(self.model(x)),axis =1)
        return {"Training Satisfaction Level":self.train_sat(self.axioms(x, y,self.p)),"Accuracy":self.train_accu(y,pred)}

    @tf.function
    def test_step(self,data ):
        x, y = data
        pred = tf.argmax(tf.nn.softmax(self.model(x)),axis =1)
        return {"Satisfaction Level" :self.test_sat(self.axioms(x, y,self.p)),"Accuracy":self.test_accu(y,pred)}

from sklearn.metrics import accuracy_score, f1_score, matthews_corrcoef, roc_auc_score
from sklearn.preprocessing import StandardScaler
import tensorflow as tf
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import time
import gc

BATCH_SIZE = 128
SEED = 42
EPOCHS = 100

X_train = np.concatenate([np.vstack(df_final[df_final['splits']=='train'][col]) for col in FEAT_NAME],axis = 1)
y_train = df_final[df_final['splits']=='train']['target']

X_test = np.concatenate([np.vstack(df_final[df_final['splits']=='test'][col]) for col in FEAT_NAME],axis = 1)
y_test = df_final[df_final['splits']=='test']['target']

X_train = np.nan_to_num(X_train)
X_test = np.nan_to_num(X_test)

y_train = np.asarray(y_train).flatten()
y_test = np.asarray(y_test).flatten()

print(X_train.shape,y_train.shape)

ds_train = tf.data.Dataset.from_tensor_slices((X_train,y_train)).batch(BATCH_SIZE).shuffle(512,seed= SEED)
ds_test = tf.data.Dataset.from_tensor_slices((X_test,y_test)).batch(BATCH_SIZE)

!rm -rf tuner_dir

# GridSearch, RandomSearch
tuner = kt.GridSearch(
    build_model,
    objective="val_Accuracy",
    max_trials=10,
    executions_per_trial=1,
    overwrite=True,
    directory="tuner_dir",
    project_name="mlp",
)

# Step 4: Search best hyperparameters
tuner.search(ds_train, epochs=EPOCHS, validation_data=ds_test, verbose=1)

tuner.get_best_hyperparameters()[0].values

tuner.results_summary()
